refactor(gui-app): name the discovery signal for what it reads
The narration input and its hook were named `discoverySettled` /
`useHostDiscoverySettled`, and their docs described "a delivered registry
listing, empty or not". That is `hasSettledFleet()`'s semantic, and it is
the one deliberately NOT used here: waiting for a delivery means waiting
out any outage the registry cannot be reached in. What they actually read
is `hasConcludedDiscovery()` - an attempt finished, delivered or failed.

A name that says "settled" invites the next reader back into exactly that
lockout, so both are renamed to `discoveryConcluded` /
`useHostDiscoveryConcluded` and the docs now state the contract they have:
an attempt has concluded for the current identity - a listing delivered
(empty or not) or a fetch that failed - false until then, and again after
a `signed-out` outcome withdraws it.

Also trims the surrounding comments to invariants. They had accumulated
change-narration - what reached the screen, for how long, what a reviewer
had caught - which describes an episode rather than the code. What each
comment owes its reader is why ∅ is ambiguous before a conclusion, why the
gate's cover fills the frame the narrator leaves, and why the arm sits
outside the pre-serve grace and is not gated on `hasBeenServed`.

No behaviour change: identifiers, comments and the file name only.
