+++ clients/gui-app/src/components/layout/host-readiness-controller.tsx
@@ -48,7 +48,7 @@ import { useRemoteSessionsPollReadiness } from "@/hooks/host/use-remote-sessions
 import { useHostBinding } from "@/lib/host";
 import { resolveAppWideHostClient } from "@/lib/host/binding-host-client";
 import { useEffectiveHostId } from "@/hooks/host/use-effective-host-id";
-import { useHostDiscoverySettled } from "@/hooks/host/use-host-discovery-settled";
+import { useHostDiscoveryConcluded } from "@/hooks/host/use-host-discovery-concluded";
 import { useHostLeases } from "@/hooks/host/use-host-lease";
 import { useSelectionAuthorityAttached } from "@/hooks/host/use-selection-authority-attached";
 import { windowNarrationAwaitsDiscovery } from "@/lib/host/window-narration";
@@ -567,26 +567,23 @@ export function DefaultHostReadyGate(props: {
 }
 
 /**
- * The narrator-owned slot's cover for the ATTACH gap. The window narrator is
- * structurally silent until the selection kernel attaches
- * (`deriveWindowNarration` returns silent on `attached: false`), and this
- * frame used to render nothing there - a blank page with only the header for
- * the whole attach latency, under a data attribute claiming a narrator that
- * was provably not rendering yet. One speaker at every moment: this card
- * shows only while the narrator cannot speak, and yields the instant it can.
+ * The narrator-owned slot's cover for the window in which the narrator cannot
+ * speak. One speaker at every moment: this card shows only then, and yields the
+ * instant the narrator can.
  *
  * The line is deliberately NOT from the F19 lane table - no lane is known to
  * be running yet; this is the window finding its authority, and claiming
  * "Starting local Traycer Host…" here would name a machine nothing has
  * resolved.
  *
- * TWO WAYS THE NARRATOR CANNOT SPEAK, and the second is why this card reads
- * more than `attached`. On a shell with no local host, an ∅ derived before
- * discovery has answered is a vacuum rather than a verdict, and the narrator
- * holds its tongue through it (`windowNarrationAwaitsDiscovery`, whose
- * definition this shares rather than restates - a second copy is how a frame
- * ends up with two cards or with none). Reading only `attached` would leave
- * that window as a header over an empty page.
+ * TWO WAYS THE NARRATOR CANNOT SPEAK, which is why this reads more than
+ * `attached`. It is structurally silent before the selection kernel attaches
+ * (`deriveWindowNarration` returns silent on `attached: false`); and on a shell
+ * with no local host it holds its tongue over an ∅ that no discovery attempt
+ * has concluded for, which is a vacuum rather than a verdict. The second
+ * condition is SHARED rather than restated (`windowNarrationAwaitsDiscovery`) -
+ * a second copy of "is the narrator able to speak" is how a frame ends up with
+ * two cards or with none.
  */
 function AttachPendingCard(props: {
   readonly presentation: DefaultHostReadinessPresentation;
@@ -597,12 +594,12 @@ function AttachPendingCard(props: {
   // outside a provider there is no local host to expect, which is also the
   // correct answer.
   const localHostExpected = useRunnerHostOrNull()?.hasLocalHost ?? false;
-  const discoverySettled = useHostDiscoverySettled();
+  const discoveryConcluded = useHostDiscoveryConcluded();
   const awaitingDiscovery = windowNarrationAwaitsDiscovery({
     attached,
     effectiveHostId,
     localHostExpected,
-    discoverySettled,
+    discoveryConcluded,
   });
   if (attached && !awaitingDiscovery) return null;
   return (
